00._Assignments/01._Data_Parser/01._Part_1/02._python/pyproject.toml
[project]
name = "02-python"
version = "0.1.0"
description = ""
authors = [
    {name = "Samim Saidi",email = "sami0880@stud.kea.dk"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
packages = [{include = "02"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
